+++ src/index.js
@@ -10,7 +10,11 @@ const formatChildren = (children) => children.length > 1 ? children : children[0
 
 export default (component, props, ...children) => {
     if (isPureSFComponent(component)) {
-        return component({ ...component.defaultProps, ...props, children: formatChildren(children) }, {});
+        const combinedProps = { ...component.defaultProps, ...props };
+        if (children.length > 0) {
+            combinedProps.children = formatChildren(children);
+        }
+        return component(combinedProps, {});
     }
 
     return React.createElement(component, props, ...children);